[CIFuzz][ClusterFuzzLite] Implement batch fuzzing (#5963)
 [ClusterFuzzLite] Support use of github actions as a filestore

+++ infra/cifuzz/fuzz_target.py
@@ -101,10 +101,9 @@ class FuzzTarget:
     # If corpus can be downloaded use it for fuzzing.
     self.latest_corpus_path = self.clusterfuzz_deployment.download_corpus(
         self.target_name, self.out_dir)
-    if self.latest_corpus_path:
-      command += docker.get_args_mapping_host_path_to_container(
-          self.latest_corpus_path)
-      command += ['-e', 'CORPUS_DIR=' + self.latest_corpus_path]
+    command += docker.get_args_mapping_host_path_to_container(
+        self.latest_corpus_path)
+    command += ['-e', 'CORPUS_DIR=' + self.latest_corpus_path]
 
     command += [
         '-e', 'RUN_FUZZER_MODE=interactive', docker.BASE_RUNNER_TAG, 'bash',
@@ -149,14 +148,14 @@ class FuzzTarget:
     # We found a bug but we won't report it.
     return FuzzResult(None, None, self.latest_corpus_path)
 
-  def free_disk_if_needed(self):
+  def free_disk_if_needed(self, delete_fuzz_target=True):
     """Deletes things that are no longer needed from fuzzing this fuzz target to
     save disk space if needed."""
     if not self.config.low_disk_space:
+      logging.info('Not freeing disk space after running fuzz target.')
       return
-    logging.info(
-        'Deleting corpus, seed corpus and fuzz target of %s to save disk.',
-        self.target_name)
+    logging.info('Deleting corpus and seed corpus of %s to save disk.',
+                 self.target_name)
 
     # Delete the seed corpus, corpus, and fuzz target.
     if self.latest_corpus_path and os.path.exists(self.latest_corpus_path):
@@ -164,10 +163,13 @@ class FuzzTarget:
       # https://github.com/google/oss-fuzz/issues/5383.
       shutil.rmtree(self.latest_corpus_path, ignore_errors=True)
 
-    os.remove(self.target_path)
     target_seed_corpus_path = self.target_path + '_seed_corpus.zip'
     if os.path.exists(target_seed_corpus_path):
       os.remove(target_seed_corpus_path)
+
+    if delete_fuzz_target:
+      logging.info('Deleting fuzz target: %s.', self.target_name)
+      os.remove(self.target_path)
     logging.info('Done deleting.')
 
   def is_reproducible(self, testcase, target_path):
